refactor(park-keeper): MiniMap 千行檔按 S0 純搬移模式拆分
- miniMapShared 承載型別、常數、幾何純函式與 marker 圖示工廠
- MiniMapController 承載 MapController、DraggableMarker、CarPositionReader
- MiniMapPhotoOverlay 承載可拖曳照片 overlay
- MiniMap 本體 946 行降至 337 行；行為零變更、既有 60 項測試斷言零刪除原樣通過

測試：typecheck、vitest MiniMap 四測試檔 60 passed（未修改任何斷言）
